Separate RefFilterOptions to own file
